Backend/FastAPI/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
bcrypt>=4.0.0
pycryptodome==3.20.0
cryptography>=41.0.0
pytest>=7.3.1
httpx>=0.24.0  # For testing
